requirements.txt
pydantic-ai>=0.0.1
pydantic>=2.0.0
httpx>=0.27.0
ollama>=0.3.0
google-genai>=0.2.0
pyyaml>=6.0
tomli>=2.0.0; python_version<"3.11"
tomli-w>=1.0.0
